--- modules/sagemaker/sagemaker-model-package-event/docs/_static/.$architecture.drawio.bkp
@@ -1,1 +1,0 @@
-<mxfile host="Electron" modified="2024-03-16T23:46:40.240Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.2.3 Chrome/102.0.5005.167 Electron/19.0.11 Safari/537.36" etag="VBlW5IeVN6Cjcos3KLfY" version="20.2.3" type="device"><diagram id="nOkpBZcZqbjnXywVXBLF" name="Page-1">7VttT+M4EP41le4+UNlxk5SPfQF2Jbhjl9Ot9r5UJnHTLGlcOQ6U+/VrJ3bbxG4bKLQFtSCIx2/JzPOMZxy3hQbT+RXDs8kNDUnSckA4b6Fhy3Ggc+6Kf1LyXEq6EJaCiMWharQU3MX/EyUESprHIckqDTmlCY9nVWFA05QEvCLDjNGnarMxTaqzznBEDMFdgBNT+iMO+UQ9hQuW8i8kjiZ6ZghUzRTrxkqQTXBIn1ZE6KKFBoxSXl5N5wOSSOVpvZT9LtfULm6MkZQ36fDrWxTc9h7HPdTvXnshn+fg55ka5REnuXrgfzCLCBey3o87+TcIaC4mKB+BP2u9zGic8kK3bl/8iqkHoOWKmoEstR23JqiX/aoAmiU5RlVQL/tVAawPD2vzw/oNrgiMUmV4UJsfrNyg+EV9mvMkTslggUIghBHDYSysM6AJZUKW0lRorz/h00SUoLh8msSc3M1wILX6JBgkZGOacsUD6OiyUrwcVSCdYzEXU2MUliDs4pGUBinbJAmeZfH9ohcjQc6y+JF8J1k5uJQKTM7k9XQeSfq28VPWaUeM5rPi9r+Kuay1I3E5ChKahyOccDkQZ/SB6AdtOUj8XEoc9sdxktQU8EgYjwXFekkcyfE5ldNhVUrIuBhRaCVOo+uiNERAacI2RYizCQnVI5m00BgXs5L5ikjR5IrQKeHsWTRRtUhTVvksTfanpQOAWjZZIX9HOy2snE60GHrJS3GhqGmn6RwMb6eh/23wH/4ySP7++Vc0+PfM6RoEJKHwU6pIGZ/QiKY4uVhK+8JQabjQy7LNNZX6LrDzi3D+rMCGc06r6CznlBOtdTmbnEpGcxaQDU8Fu3YzMJJgLqBadeYWnaqut5IBS/P5bsV6CNSswgsPpzrVDLO4i9fbCnYNn1pwU4ju88ywY/ZAeDDRRrK5kXWwX+NeVgknmvcB6HaGBkVV4wp1NC+v8T1JbmkW87jg/z3lnE63Ejcg0g9VIbTNmeFsVj7oOJ7L+zC8GVznpYhscS9ihoiMQjLGecJHhWwktDxiRKOvv4b9TXG81ku4NZh1nLZ3DpYf13AaHYvPcPy2Bszbew3HQOLX3o0QfKfCAHsH4iXoQu/zAZHRZI8w64A2WP2gZihz2q73XijzP+Pa5HRUslGGwxsaejsuYrspv2NQ/Ht+CHIf9SrTnMtCdyMkexAWCwsRdleGoZXg8t1J7qBuI1p33i3e9AxYDURqfxvPiETOZnhZU0NremhLEa1popkqVpoVyZtlhrrQJvNNITSb6XzPFNpktuS23htaesNa7/WpZVO2irrO0BeVK3XDWCSDinSpdL0mncWn0z+35XXj4nMEnDYWbDuhVRxYZrIiC85sOW0gsD3T2LaSX/jZvbEfOma+uVf2m5tCd+VyedoUOm0KfeBNoS3M3UzSztY9Icfb656Q3kg/urh7k0/ZGnejhmE3BHZr7SfsRoaHNGxhgbjBB+SdO/7FWvDXSV6DutvzBl2vKbrXW2Ut5GvboBDAtmMGpotmq7CH9U23N1O9ay5OOCI3+EH4PgfoF1C3OHiQb3YccFU4rlPA+oEC1l7fOe+9LGCF0BsK4+0csE7jMExsy9GiorYiHShizQS4pwXmUT+pPQMraVhbLJ3DOwrX9BNa9uZuwkxgtWswl8333hwBnn/pdY91c+T1O58LEI6mhW4XkHtJ1PUKuNVCMYiscNtryuSvhZsZGZzg9rHgBuHx4c18x6jxhk54++h46x4f3qDpxj5BsqmJvT3bdHfMNu0nFSCqWho5S0vv67QCMPMpjrnMnIIJTiNhPscrdnPuRX7lRfLqj4I7f+7d0bzBS6bm8f0BTjM0juVdYDgjy5412uOa6HQO76MWB72WUL4qTjLeEI5DzPGndGFNN8zQRnuCNvI9r2LSM73F+TIn12MMP680UJsrjX2gC2tAKAesddaj0/E4I+/jGM09vtPhmYMennmNm/Kq4Dq+0zPQ3M88HeB4gwMc+1r4EPIOv/ABAy7HsbI1XsTcpnH4rieGd9OzuST0siyfmmz9WOrXb6+2q39bEOF5yK8GEfot5WszpV0WelFcfr2mbL78khK6+A0=</diagram></mxfile>
