--- modules/sagemaker/sagemaker-model-package-event/docs/_static/architecture.drawio
@@ -1,1 +1,0 @@
-<mxfile host="Electron" modified="2024-03-16T23:48:01.709Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.2.3 Chrome/102.0.5005.167 Electron/19.0.11 Safari/537.36" etag="-1_E-aS3OTB5lrjLz8IR" version="20.2.3" type="device"><diagram id="nOkpBZcZqbjnXywVXBLF" name="Page-1">7Vttb+I4EP41SHcfiuw4CfCRl7a7Uveuuz3dau8LchMTsg0xcpyW3q9fO7GBxAaypQVaQas2Hr8lM88znnFMCw1ni2uG59MvNCRJywHhooVGLceBTs8T/6TkuZR0ISwFEYtD1WgluIv/J0oIlDSPQ5JVGnJKEx7Pq8KApikJeEWGGaNP1WYTmlRnneOIGIK7ACem9Hsc8ql6Cg+s5J9IHE31zBComhnWjZUgm+KQPq2J0GULDRmlvLyaLYYkkcrTein7XW2oXd4YIylv0uHn1yi47T9O+mjQvfFDvsjBjws1yiNOcvXA/2AWES5k/e938m8Q0FxMUD4Cf9Z6mdM45YVuvYH4FVMPQcsTNUNZajteTVAvd6oCaJbkGFVBvdypCmB9eFibH9ZvcE1glCrDg9r8YO0GxS8a0JwncUqGSxQCIYwYDmNhnSFNKBOylKZCe4MpnyWiBMXl0zTm5G6OA6nVJ8EgIZvQlCseQEeXleLlqALpHIu5mBqjsARhl4+kNEjZJknwPIvvl70YCXKWxY/kG8nKwaVUYHIur2eLSNK3jZ8ytx0xms+L2/8s5rLWjsXlOEhoHo5xwuVAnNEHoh+05SDxcyVxOJjESVJTwCNhPBYU6ydxJMfnVE6HVSkhk2JEoZU4jW6K0ggBpQnbFCHOpiRUj2TSQmNczEoWayJFk2tCZ4SzZ9FE1SJNWeWzNNmfVg4Aatl0jfyudlpYOZ1oOfSKl+JCUdNO0wUY3c7Cztfhf/jTMPn7x1/R8N8Lp2sQkITCT6kiZXxKI5ri5HIlHQhDpeFSL6s2N1Tqu8DOT8L5swIbzjmtorOcU0600eVscyoZzVlAtjwV7NrNwEiCuYBq1ZlbdKq63koGrMzX8SrWQ6BmFV54ONWpZpjlXbzcVrBr+NSCm0J0n2eGHbMHwoOpNpLNjWyC/Qb3sk440XwAQNcdGRRVjSvU0by8wfckuaVZzOOC//eUczrbSdyASD9UhdAuZ4azefmgk3gh78PwZnCTlyKyxb2IGSIyDskE5wkfF7Kx0PKYEY2+wQb2N8XxRi/h1WDmOm2/B1Yfz3AarsVnOJ22Bszrew3HQOLn/hch+EaFAQ4OxCvQhf7HAyKjyQFh5oI2WP+gZihz2p7/VijrfMS1yXFVslGGw1sa+nsuYvsp3zUo/i0/BrlPepVpzmWhuzGSPQiLhYUIuyvD0Epw+eYkd1C3Ea3dN4s3fQNWQ5Ha38ZzIpGzHV7W1NCaHtpSRGuaaKaKlWZF8maZoS60yTqmEJrNdL5nCm0yW3Jb7w0tvWGt9+bUsilbRZ076ojKtbpRLJJBRbpUul6TzuLjDnq2vG5SfE6A08aCbSe0igPLTFZkwZktpw0Etuca21byCz97MPZDx8w3D8p+c1Porlwuz5tC502hd7wptIO520nq7twTcvyD7gnpjfSTi7u3+ZSdcTdqGHZDYLfWYcJuZHhIwxYWiBt8QH7P6VxuBH+d5DWoe31/2PWbonuzVTZCvrYNCgFsO2Zgumy2DntY33R7NdV75uKEI/IFPwjf5wD9AuoWBw/yzY4DrgvHdQ5Y31HA2h84vf7vBawQ+iNhvL0D1lkcholtOVpW1FakI0WsmQD3rMA8GiS1Z2AlDWuLpXN8R+GZfkLLXt1NmAmsdg3msvnWmyPA71z53VPdHHn5zucShONZodsl5H4n6noB3GqhGERWuB00ZepshJsZGZzh9r7gBuHp4c18x6jxhs54e+94654e3qDpxj5AsqmJvTvb9PbMNu0nFSCqWho5K0sf6rQCMPMpjrnMnIIpTiNhPscvdnPuRX7lR/Lqj4I7fx7c0bzCS6bm8f0RTjM0juU9YDgjy541OuCa6LjH91HLg14rKF8XJxnF8DjEHH9IF9Z0wwxttSdoo54HKya90Ib6PSfXZww/rzVQmyuNfaC6ixUQygFrnfXodDLJyNs4RnOP73x45qiHZ17iprwquE7v9Aw09zPPBzhe4QDHoRY+hPzjL3zAgMtprGyNFzGvaRy+74nh/fRsLgn9LMtnJlvfl/r126vd6t8VRHhdp1cNIvRbypdmSvss9KK4+npN2Xz1JSV0+Qs=</diagram></mxfile>
