--- modules/sagemaker/sagemaker-model-package-event/tests/test_app.py
@@ -1,26 +1,0 @@
-import os
-import sys
-
-import pytest
-
-
-@pytest.fixture(scope="function")
-def stack_defaults():
-    os.environ["SEEDFARMER_PROJECT_NAME"] = "test-project"
-    os.environ["SEEDFARMER_DEPLOYMENT_NAME"] = "test-deployment"
-    os.environ["SEEDFARMER_MODULE_NAME"] = "test-module"
-
-    os.environ["CDK_DEFAULT_ACCOUNT"] = "111111111111"
-    os.environ["CDK_DEFAULT_REGION"] = "us-east-1"
-
-    os.environ["SEEDFARMER_PARAMETER_model_package_group_name"] = "dummy123"
-    os.environ["SEEDFARMER_PARAMETER_target_event_bus_name"] = "dummy321"
-    os.environ["SEEDFARMER_PARAMETER_target_account_id"] = "dummy321"
-
-    # Unload the app import so that subsequent tests don't reuse
-    if "app" in sys.modules:
-        del sys.modules["app"]
-
-
-def test_app(stack_defaults):
-    import app  # noqa: F401
