+++ modules/sagemaker/sagemaker-model-package-group/docs/_static/.$architecture.drawio.bkp
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2024-03-23T01:38:58.537Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.2.3 Chrome/102.0.5005.167 Electron/19.0.11 Safari/537.36" etag="JrvHVmd0AZ8bNjwvECL3" version="20.2.3" type="device"><diagram id="nOkpBZcZqbjnXywVXBLF" name="Page-1">7Vpbc6M2FP41nmkf6kFgwH70Jcm2TdrsptNt+uJRQMZsBPIIEdv763sEAhskEje7dpI2TiZBn66c850buOdMk80Fx6vlFQsJ7dlWuOk5s55to4HtwT+JbEtk6NolEPE4VIN2wE38lSjQUmgehyRrDBSMURGvmmDA0pQEooFhztm6OWzBaHPXFY6IBtwEmOro5zgUy+ourB3+gcTRstoZWaonwdVgBWRLHLL1HuSc9ZwpZ0yUV8lmSqgUXiWXct55R299ME5ScciELx+j4Hr8sBg7k+GlF4pNbt3+pFZ5wDRXN/wH5hERgI0/38i/QcBy2KC8BbGt5LJicSoK2boT+IWtp1bPhZ6pbPVttwW0234TQHpLrtEE2m2/CaD28qi1P2ofcA/QWo3lrdb+1t4B4deZsFzQOCXTmoUWgBHHYQzamTLKOGApS0F6k6VIKLQQXK6XsSA3KxxIqa7BggBbsFQoO0B21VaCl6sC0wWGvbhao9AE4WcPpFRIOYZSvMriu3oWJ0HOs/iBfCJZubhEgZMreZ1sImm+fbzOBv2Is3xVHP9n2MvYO4fLeUBZHs4xFXIhwdk9qW60Zzvwcy55OFnElLYE8EC4iMHExjSO5PqCye2walGyKFYEqcRpdFm0Zo6lJGHaIsTZkoTqlnSzqDgOu5LNHqTM5IKwhAi+hSGqd1A5H+WzKmNf7xwAqrDlnvE7AwVi5XSieumdXcKFMk2zmQa/4l+S5f0kvL9dX3ljMl/QrwYzLdQN0F2eabaZ3RMRLJU8jMzskmQHY/d1CMMnljUczDStq8ENbVSqvsR3hF6zLBZxQak7JgRLnuRCQCS1mzbzlH3gbFXe6CLeyHNoBoK6iE/kiDsIQxGZh2SBcyrmBTYHKc85yVjOpaVOOgjVpF23v+0knus2eGf7Xh+4WH+QRsOBgYW230fDb+fhxppdJ6H/cfo3/jClv9/+Fk3/NPDwppTJe7h4Dxf/23DxZLRwjhYtzFaKNPsj4NMqVjAulixiKaZnO3QCekrDWiy7MZdMirugzhcixFZxDeeCNclJNrH4S04H4pet272e2UatXDS2ndJ/zO1U/rdznKMKhDKFfWQ9JXkplUc1zAnFAqygWVYY9KWmXkvj2jHDtux+06WjUUvp5T2peS291wd5PhUczWFr1DBYnGaejjey/bNOW2z7nJbluWNvOvT+tboPt0C7aYGOboH2yGCByDmWBbp6nIRC8Qrfgxu2rapKvsbBvSw/beui8KGP5nHGCGqMoqZIaoymekRtDCtinGGHNmjCfB1E+rAqLOqgCTPlAO3ZyDAbtWZ3R+BD02LoG4wn9mi81zeLIWaq7DaV7lPPmxHyZqA8g30tis+hyXMShyE1Rca6oxUcvyV7Nsb8KhcuIz5kC5kp9mdA7qTgvDOhrXvgpRm24rb9sk7CNWTT7pFchKe5iMot6NH72AWd5fnn3vC/V9DVBJwnhWxruj0r/TicanVgUVwb6VwzVW5HSwj9Tq7p6cA7194Y10avi2vDTq4571x741yra9tXwjWku68TFLrfuWititGnq1b3NFUr8vxW1eo4qG+3nimWxz1a4VpJZb+EEljIYilY4jQCBdpe8SzpDkoqL5JXPxSW8+PJ3cx3eB5+eEr/Ao/ED07fXUtzRSvCY+AC4Tc17BwzGg5aj9Ctl3ZR+qOATzko5g1y9Iih8GA+cpDd/KQE8k9HIOOrP1vnT/UiyrauGY2D7enJdG4NkfdayfR8bwfOKomzDE6Y9Q58y9fNuk6Kec0kytMJBuHWH53QSdVfUXnhPOrglMnVU6bH3psfO2NyfLcPtYbvDgfW0HId1Mye2np7duoEzd33hcrhu29dOWf/AA==</diagram></mxfile>
