+++ modules/sagemaker/sagemaker-model-package-group/docs/_static/architecture.drawio
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2024-03-23T01:42:22.857Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.2.3 Chrome/102.0.5005.167 Electron/19.0.11 Safari/537.36" etag="Kohy90ZlEb0quRIcgclY" version="20.2.3" type="device"><diagram id="nOkpBZcZqbjnXywVXBLF" name="Page-1">7Vtbc5s4FP41ntl9qIeLDfajL0m7u+mu23S2m33xyCBjGoE8IGK7v36PQGBAwmESX9Ku004GHYQE53znOxdIx5wE2/cRWq8+UheTjqG524457RjG0NDgNxfsMkG/LwRe5LuZSN8L7v3vWAjzaYnv4rgykVFKmL+uCh0ahthhFRmKIrqpTltSUt11jTwsCe4dRGTpV99lq0w6yJ+Cyz9g31vlO+uaOBOgfLIQxCvk0k1JZN50zElEKcuOgu0EE667XC/ZdbcNZ4sbi3DI2lzw7ZPnzEZPy5E5HtxZLtsm2sM7scoTIol44C8o8jAD2ejrPf/tODSBDbJHYLtcL2vqhyzVbX8M/2Hridbpw5kJH3WNfk1QH9tVgS6P+BpVQX1sVwV6fXm9tr9ev8GSQBpVltdq+2ulG4T/5pgmjPghnhQo1EDoRcj1wToTSmgEspCGoL3xigUERjocblY+w/dr5HCtbsCBQLakIRN+oBv5WCierwpIZwj2isQaqSVwdPOEM4NkcwhB69hfFFdF2Emi2H/Cn3GcLc6lgMk1Pw62HvfeLtrEva4X0WSd3v5vsJfy7BwO5w6hiTtHhPGFWEQfcf6gHcOEf7cch+OlT0hNAU84Yj642Ij4Hl+fUb4dEiOCl+mKoBU/9O7S0dTUhCZUW7goXmFXPJLsFjnGYVe8LYmEm7zHNMAs2sEUcbaXk0/OWWK42ROAnstWJec3e0KIBOl4xdJ7v4QD4ZpqN3X+QL8Hq8ex+/iw+WiN8HxJvr/rSf6HXaApMaQRW1GPhojc7KVjsFPoFmrZz7mjXN0pdL5hxnYCayhhtArObE++USPjHOKUmCaRgw88VIMRIkwQA6BWqVyhUXHpjON/bzzu+SXbGT27ugJL+U1cVDNLcRcvt5RMqKljgmiRxJIV40fMnFVuIhWHNGG+gVvK3gbTx5o26E0l/xSTK36TO+UdWmAyo7HP/NT5F5QxGjzrtQ7mJFQF0HNMhuJ19qBLf8vvQ6IyvYmiMJ+xgITBw3MXL1FC2DyVzUHL8wjn2BuvceSDSXF0n9EJLNQroC0xQltwNzJHvwY+2+oCmRQ/ek8ikp6CRwy7qw9OxCRgiJ+QSkyRGWa5S/M865WU8yrVmxI1fE5AvVdOKHNCawqIQHdz8+Iurg/k5EDl0ydLDayGDF4ONopMTErb+iNrMrDKONAbzVw3XQ2WxVKnskSLLM1UWKKo0I5uClsyxR0KFi46u4tPe1Ci2G/VxV8e9kmqzvkyCZ3sNk+GLbOKrUs7+VBClowpobAN2qX65Dosw0xUfVaDvqv1/FRduEW8dkSlopIyxErjgooLZMz2knEcovUXmmXe5bTg+Naz35b1cqK6mq+d+QZvzHxyTXc13wHzDS9mvq02nQWu/WnyL/owIX89/OlN/lZU5PdZOXFtcV5bnP/bFuezubN5sg6n2kt1yf/O0JfAW5/9wy8H4Gejh9KZ6VasnA52jdo/RDvl1oVynqJ1oV5PaL7WRDl6/9TQjG6txB3WjJ490zFaqIc0cijctillTWto2DdHql7bmru9BxpVDzRlDywCaaV6NU/lgX05TiIPf0SPQMOGlr/YnSHnkb8yNbT3KYceLG2VEVQZRVWRVBlN5YhamZbGOMUOdaFKZstCXZ6Wh0VZqJKpcoD61briar12dXMEbtspgHO90dgYjkrnpj7ETFHwh5w+5VaCrltTMJ7Cv5bpT9t+QuC7LlFFxuJELTi+pqGgbhiKtwJZxOfZtCr2xwDuIMW8OSa1Z4gyN6zFbeOyJNGXOSKXHZ0i5F5jTgty9D51j0uz7Ftr8PP1uAoAzoNUtwXcXpR+tIeaXmt5DS9ctsnN1Bxriur7irUfC2vDt4W1QSPWzCvWfnCsFbXtG8HaJV/AH61ozYvR56vW/nmqVt2ya1Wraepdo/YVxfG+/TmolXIJxRDjxZKzQqEHBjSstJe0gJLK8vjRL6nn/Hp2mjnCVwDtU/oLfBzUOn3vaxIVHfyi4ATRsFf7aki7NEXJrYDrlyqn/lLlDLiyz4cr5Ys0Q4ZV/qWeoc0o8Z3d+TF2qw10661i7OUkCFgL/DiGO4w7LV/1NaOuEWJWNbeyZIBBFLaHZ8RY/3LZ1cHPG1tnVn05szrJ19MNWobh/k9SssRo/3c95s1/</diagram></mxfile>
